requirements.txt
drf-composable-permissions
django-filter
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
python-decouple>=3.8
pytest-django>=4.8.0
drf-spectacular>=0.27.0
stripe